Split JS in to smaller module files
