+++ assets/src/scripts/_signals.js
@@ -1,0 +1,6 @@
+/* eslint-disable import/prefer-default-export */
+import { signal } from "@preact/signals";
+
+const openFile = signal();
+
+export { openFile };
